Extract navigation view helpers into own namespace
This commit makes use of Form::Routing’s list of form
names instead of duplicating the list.
